Integrate UZDoom 5.0.0-pre (emileb uz_5.0_pre) — builds & runs on device
Swaps the native engine from GZDoom 4.15 to emileb's uz_5.0_pre (UZDoom
5.0.0-pre, a GZDoom fork). Verified booting doom2.wad on an arm64 device:
"UZDoom version 5.0.0-pre", base pk3s + IWAD load, OpenAL+EFX up, DOOM II
menu renders. Both ABIs (arm64-v8a, armeabi-v7a) built.

Engine:
- Re-vendored uz_5.0_pre over doom/src/main/jni/engine/gzdoom/ (C++20 +
  abseil come from the engine's own mobile/Android_*.mk).

Build-graph fixes to fit the existing support libs (mobile/Android_src.mk,
ZMusic/source/Android_zmusic.mk):
- openal-soft -> openal (reuse vendored module + include path)
- drop core_shared (absent; 4.15 also linked without it)
- webp include path -> libraries/webp/include
- ZMusic: link fluidsynth-static (static) instead of shared fluidsynth

Compile fixes:
- gitinfo.h stub: add GIT_TAG / GIT_DISTANCE (now required by version.h)
- vendored OpenAL al.h/alc.h: define AL_API_NOEXCEPT(17)/ALC_API_NOEXCEPT(17)
  as no-ops — UZDoom's newer thirdparty alext.h/efx.h annotate prototypes
  with them and the older header wins the AL_AL_H guard on the !DYN path

Re-applied engine patches for 5.0:
- crashcatcher disabled on Android (fork-debugger deadlock), i_main.cpp
- ES 3.1 -> 3.0 GL context, sdlglvideo.cpp

App/assets:
- Game.kt getLibraries(): gzdoom->uzdoom, zmusic_g4d->zmusic_uz
- Base data rebuilt from wadsrc* (zip): uzdoom.pk3, uzdoom_game_support.pk3
  (+ widescreen/lights/brightmaps); removed gzdoom_dev_gl3.pk3/game_support.pk3
- LaunchState/EditorLauncher unpack uzdoom.pk3 + uzdoom_game_support.pk3 into
  res/ (BASEWAD/OPTIONALWAD); IwadScan exclusions updated

The Clibs_OpenTouch/idtech1 JNI glue compiled against the 5.0 engine unchanged
(its -DUZDOOM path already matched).

+++ doom/src/main/jni/engine/gzdoom/src/common/rendering/gles/glad/include/KHR/khrplatform.h
@@ -241,9 +241,9 @@ typedef unsigned long long int khronos_uintptr_t;
 typedef signed   long long int khronos_ssize_t;
 typedef unsigned long long int khronos_usize_t;
 #else
-typedef signed   long  int     khronos_intptr_t;
+typedef signed   int     khronos_intptr_t;
 typedef unsigned long  int     khronos_uintptr_t;
-typedef signed   long  int     khronos_ssize_t;
+typedef signed   int     khronos_ssize_t;
 typedef unsigned long  int     khronos_usize_t;
 #endif
